requirements/base.txt
python3-ldap==0.9.5.4
PyYAML==3.11
pyzabbix==0.7.1
